refactor(web): extract demo fixtures
